+++ .eslintignore
@@ -2,5 +2,5 @@
 /docs/**/*
 /ts*/**/*
 /types/**/*
-.eslintrc.js
+.eslintrc.cjs
 rollup.config.js
