+++ package.json
@@ -1,30 +1,26 @@
 {
   "name": "@0b5vr/tweakpane-plugin-profiler",
-  "version": "0.3.0",
+  "version": "1.0.0",
   "description": "Profiler plugin for Tweakpane",
   "main": "dist/tweakpane-plugin-profiler.js",
-  "types": "types/index.d.ts",
+  "type": "module",
+  "types": "dist/types/index.d.ts",
   "author": "0b5vr",
   "license": "MIT",
   "files": [
-    "/dist/",
-    "/ts*/",
-    "/types/",
-    "/LICENSE",
-    "/package.json",
-    "/README.md"
+    "dist"
   ],
   "scripts": {
     "prepare": "run-s clean build",
-    "start": "npm run watch",
-    "assets": "run-s clean build assets:version assets:zip",
-    "assets:version": "node scripts/assets-append-version.js",
-    "assets:zip": "zip -x \"*types*\" -j -r $(cat package.json | npx json name)-$(cat package.json | npx json version).zip dist",
-    "clean": "rimraf dist/ types/ ts*/ *.tgz *.zip",
+    "prepublishOnly": "npm test",
+    "start": "run-p watch server",
+    "test": "eslint --ext .ts \"src/**/*.ts\"",
+    "clean": "rimraf dist",
     "build": "run-p build:*",
     "build:dev": "rollup --config rollup.config.js",
+    "build:dts": "tsc --project src/tsconfig-dts.json",
     "build:prod": "rollup --config rollup.config.js --environment BUILD:production",
-    "build:types": "tsc --declaration --declarationDir ./types --emitDeclarationOnly && downlevel-dts types ts3.4/types",
+    "server": "http-server -c-1 -o /index.html",
     "watch": "run-p watch:*",
     "watch:sass": "onchange --initial --kill \"src/sass/**/*.scss\" -- npm run build:dev",
     "watch:ts": "onchange --initial --kill \"src/**/*.ts\" -- rollup --config rollup.config.js"
@@ -34,25 +30,25 @@
     "@rollup/plugin-node-resolve": "^13.0.0",
     "@rollup/plugin-replace": "^2.4.1",
     "@rollup/plugin-typescript": "^8.2.0",
-    "@tweakpane/core": "^1.0.6",
-    "@typescript-eslint/eslint-plugin": "^4.15.2",
-    "@typescript-eslint/parser": "^4.15.2",
+    "@tweakpane/core": "^2.0.1",
+    "@typescript-eslint/eslint-plugin": "^5.62.0",
+    "@typescript-eslint/parser": "^5.62.0",
     "autoprefixer": "^10.2.4",
-    "downlevel-dts": "^0.7.0",
-    "eslint": "^7.20.0",
     "eslint-plugin-sort-imports-es6-autofix": "^0.6.0",
-    "node-sass": "^6.0.0",
+    "eslint": "^8.46.0",
+    "http-server": "^14.1.1",
     "npm-run-all": "^4.1.5",
     "onchange": "^7.1.0",
     "postcss": "^8.2.6",
     "rimraf": "^3.0.2",
-    "rollup": "^2.39.1",
     "rollup-plugin-cleanup": "^3.2.1",
     "rollup-plugin-terser": "^7.0.2",
-    "tweakpane": "^3.0.5",
-    "typescript": "^4.1.5"
+    "rollup": "^2.39.1",
+    "sass": "^1.49.9",
+    "tweakpane": "^4.0.0",
+    "typescript": "^4.9.5"
   },
   "peerDependencies": {
-    "tweakpane": "^3.0.5"
+    "tweakpane": "^4.0.0"
   }
 }
