+++ src/ProfilerBladeControllerConfigs.ts
@@ -1,0 +1,8 @@
+import { Blade, LabelProps } from '@tweakpane/core';
+import { ProfilerController } from './ProfilerController.js';
+
+export interface ProfilerBladeControllerConfig {
+  blade: Blade;
+  labelProps: LabelProps;
+  valueController: ProfilerController;
+}
