+++ src/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "compilerOptions": {
+    "lib": ["DOM", "ES2015"],
+    "moduleResolution": "Node16",
+    "strict": true,
+    "target": "ES6"
+  }
+}
