+++ src/utils/rgbArrayToCssString.ts
@@ -1,4 +1,4 @@
-import { saturate } from './saturate';
+import { saturate } from './saturate.js';
 
 export function rgbArrayToCssString( array: number[] ): string {
   const arrayPrepared = array.map( ( v ) => saturate( v ) * 256.0 );
