--- tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "compilerOptions": {
-    "lib": ["dom", "es2017"],
-    "moduleResolution": "node",
-    "strict": true,
-    "target": "es2017",
-    "forceConsistentCasingInFileNames": true
-  },
-  "include": ["./src"]
-}
